Core: Image now only resizes the last valid representation and removes all others. OpenGL: Added reference counting for Texture2D as well as TextureObserver in order to enable sharing of OpenGL and OpenCL textures. This should probably be done with buffers and volumes as well. OpenCL: Added converter from ImageCLGL to ImageGL and fixed sharing with OpenGL

+++ include/inviwo/core/datastructures/data.h
@@ -234,7 +234,7 @@ T* Data::getEditableRepresentation() {
 
 template<typename T>
 bool Data::hasRepresentation() const {
-    for (int i=0; i<static_cast<int>(representations_.size()); i++) {
+    for (size_t i=0; i<representations_.size(); i++) {
         T* representation = dynamic_cast<T*>(representations_[i]);
         if (representation) return true;
     }
